Create separate collision handler systems for Spaceship and Enemy entities.
